Add events file to install scripts
git-svn-id: file:///sources/wired/zanka-full-svn/branches/P7/WiredServer@7879 94223b32-352b-4d6d-a509-7b239cf53eea

+++ install.sh
@@ -32,6 +32,10 @@ fi
 
 install -m 644 "$SOURCE/Wired/etc/wired.conf" "$LIBRARY/Wired/etc/wired.conf.dist" || exit 1
 
+if [ ! -f "$LIBRARY/Wired/events" ]; then
+	install -m 644 "$SOURCE/Wired/events" "$LIBRARY/Wired" || exit 1
+fi
+
 install -m 755 -d "$LIBRARY/Wired/files/" || exit 1
 
 if [ ! -f "$LIBRARY/Wired/groups" ]; then
